+++ __tests__/getType.test.ts
@@ -1,0 +1,9 @@
+import getType from '~/getType';
+
+test.each([
+  [{hello: 'I am a good old object'}, 'Object'],
+  ['', 'String'],
+  [undefined, 'Undefined'],
+])('getType(%o) is %s', (a, expected) => {
+  expect(getType(a)).toBe(expected);
+});