+++ types/getType.d.ts
@@ -1,0 +1,8 @@
+declare const _default: (payload: any) => string;
+/**
+ * Returns the object type of the given payload
+ *
+ * @param {*} payload
+ * @returns {string}
+ */
+export default _default;